move list visual to ui components
